Fixed footer styles, refactored reused components
Fixed footer style so that footer is no longer fixed to bottom of page no matter the resolution.

Refactored signup page to avoid relying on absolute positioning (as it messed with footer styles).

Refactored footer so that it is a global component, and modified sign up page to use the global footer.