Changed restaurant dummy data to be list of rtrnts
Changed the dummy data setup so that we have a list of
restaurants and a list of users. This allows us to use the list of
restaurants similar to the way our database will store them. We
can list restaurants by iterating over all restaurants in the list,
and we can allow users to visit restaurant pages based on
restaurant id (using props.match.params).